+++ src/modules/auth/passwd/ForgetPwdScreen.js
@@ -14,7 +14,7 @@ function ForgetPwdScreen(props) {
 	return (
 		<Container>
 			<Row className="justify-content-center">
-				<Col md="5" className="mt-3">
+				<Col md="6" className="mt-3">
 					<ForgetPwdCard app={props.app} />
 				</Col>
 			</Row>
@@ -57,10 +57,7 @@ function ForgetPwdCard(props) {
 				throw new Error(t("ForgetPwdScreen|Something went wrong, can't reset the password"));
 			}
 		} catch (e) {
-			props.app.addAlert(
-				"danger",
-				t("ForgetPwdScreen|Something went wrong, can't reset the password")
-			);
+			props.app.addAlert("danger", `${t("ForgetPwdScreen|Something went wrong, can't reset the password")}. ${e?.response?.data?.message}`, 30);
 			return;
 		}
 
@@ -144,11 +141,10 @@ function ForgetPwdCard(props) {
 
 	return (
 		<Form onSubmit={handleSubmit(onSubmit)}>
-
 			<Card className="animated fadeIn auth-card">
 				<CardHeader className="border-bottom card-header-login">
 					<div className="card-header-title" >
-						<CardTitle className="text-primary" tag="h2">{!invalidCode ? t('ForgetPwdScreen|Cannot login?') : t('ForgetPwdScreen|Invalid password reset link')}</CardTitle>
+						<CardTitle className="text-primary" tag="h2">{!invalidCode ? t("ForgetPwdScreen|Can't login?") : t('ForgetPwdScreen|Invalid password reset link')}</CardTitle>
 						<CardSubtitle tag="p">
 							{!invalidCode ? t('ForgetPwdScreen|Reset password here') : t('ForgetPwdScreen|The link has expired, you can reset your password here')}
 						</CardSubtitle>
@@ -211,11 +207,7 @@ function ForgetPwdCard(props) {
 						{t("ForgetPwdScreen|Go back")}
 					</Button>
 				</CardFooter>}
-
 			</Card>
-
-
 		</Form>
 	);
 }
-