+++ src/modules/auth/utils/MessageScreen.js
@@ -11,7 +11,7 @@ function MessageScreen(props) {
 	return (
 		<Container>
 			<Row className="justify-content-center">
-				<Col md="5" className="mt-3">
+				<Col md="6" className="mt-3">
 					<MessageCard app={props.app} />
 				</Col>
 			</Row>